Fix package address of ContactTabParser and NotesTab parser
