+++ src/main/java/donnafin/logic/commands/SwitchTabCommand.java
@@ -8,7 +8,9 @@ import donnafin.logic.Logic;
 import donnafin.logic.PersonAdapter;
 import donnafin.logic.commands.exceptions.CommandException;
 import donnafin.logic.parser.AssetsTabParser;
+import donnafin.logic.parser.ContactTabParser;
 import donnafin.logic.parser.LiabilitiesTabParser;
+import donnafin.logic.parser.NotesTabParser;
 import donnafin.logic.parser.PolicyTabParser;
 import donnafin.logic.parser.exceptions.ParseException;
 import donnafin.model.Model;
