+++ src/test/java/donnafin/logic/parser/ClientViewParserContextTest.java
@@ -17,7 +17,6 @@ import org.junit.jupiter.api.Test;
 
 import donnafin.logic.PersonAdapter;
 import donnafin.logic.commands.ClearCommand;
-import donnafin.logic.commands.ContactTabParser;
 import donnafin.logic.commands.DeleteCommand;
 import donnafin.logic.commands.ExitCommand;
 import donnafin.logic.commands.FindCommand;
