+++ src/test/java/donnafin/logic/parser/ParserContextTest.java
@@ -9,7 +9,6 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import donnafin.logic.PersonAdapter;
-import donnafin.logic.commands.ContactTabParser;
 import donnafin.model.Model;
 import donnafin.model.ModelManager;
 import donnafin.model.UserPrefs;
